--- src/ewallet/ewallet_opts.d.ts
@@ -1,7 +1,0 @@
-export interface XenditOptions {
-  publicKey?: string;
-  secretKey: string;
-  xenditURL?: string;
-  eWalletURL: string;
-}
-
